Add Introduction to the Project
Add introduction paragraph to U.S. Medical Insurance Costs. The introduction contains context as to the why behind the project, and it provides placeholder spots for potential questions that the group may have and may want to answer. There is also a section for necessary libraries for the project. There is a note specifying what each library does and why it is helpful.